feat: update pp

+++ bathbot/src/active/impls/map.rs
@@ -104,11 +104,11 @@ impl MapPagination {
 
         let mut seconds_total = map.seconds_total;
         let mut seconds_drain = map.seconds_drain;
-        let mut bpm = map.bpm;
+        let mut bpm = map.bpm as f64;
 
         let clock_rate = self.mods.legacy_clock_rate();
-        seconds_total = (seconds_total as f32 / clock_rate) as u32;
-        seconds_drain = (seconds_drain as f32 / clock_rate) as u32;
+        seconds_total = (seconds_total as f64 / clock_rate) as u32;
+        seconds_drain = (seconds_drain as f64 / clock_rate) as u32;
         bpm *= clock_rate;
 
         let mut info_value = String::with_capacity(128);
@@ -222,7 +222,7 @@ impl MapPagination {
         let _ = write!(
             info_value,
             "BPM: `{}` Objects: `{}`\nCS: `{}` AR: `{}` OD: `{}` HP: `{}` Spinners: `{}`",
-            round(bpm),
+            round(bpm as f32),
             map.count_circles + map.count_sliders + map.count_spinners,
             round(map_attrs.cs as f32),
             round(map_attrs.ar as f32),
